frt/requirements.txt
# Translation Service Dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-docx==1.1.0
mammoth==1.6.0
python-dotenv==1.0.0
openai==1.3.7